requirements.txt
fastapi
pytest==8.3.2
ffmpeg-python
pydub
google-cloud-aiplatform
moviepy
gunicorn
azure-cognitiveservices-speech
elevenlabs
python-dotenv
python-multipart
uvicorn
google-cloud-storage